Revert
The last check-in won't work since the system_incl.make will be
overwritten by XPS and could possibly create version conflicts when
other users try to check-in their projects.

The system_incl.make
a) MUST NOT be checked-in to the repo and
b) MUST be generated on the fly since GUI will modify this file

This issue will be parked until we find a better solution. For now, the
*project* Makefile will be a "Makefile wrapper" to generate
system_incl.make.