Initial commit;  AXI stream simulation support, recording and stimulation using 		our previously defined AXI grammar.
Three new pcores:
	- a common support package,
	- a simple AXI stream recorder,
	- a simple AXI stream stimulator.

At this point, only AXI stream is supported.  AXI4 Lite to come.  AXI grammar
files presently have to be hand-coded, but support for translating PCAP traces
is imminent.

AXI transaction grammar is defined at:
	http://netfpga10g.pbworks.com/w/page/32284058/AXI-transaction-grammar

The intention is that support for AXI4 Lite links will be provided by forking
the relatively simple recorder and stimulator pcores, and extending the AXI
grammar as appropriate.  To that end, as much code as is practicable has been
put into the support package.