Initial commit: Automatic MHS parser and substitutor for simulation.
Takes an MHS file, and replaces applicable pcore instances with instances of
nf10_axis_sim_{stim,record}.  Includes an 'undo' feature to reverse the changes
made.

Not extensively tested as yet, but it seems to work in the loopback project.
I expect problems with TDATA/TSTRB widths.  Probably the easiest way to solve
this is to change the default widths in the pcores themselves, and then add
code to look for any overrides from the default - assuming that everybody uses
default widths of 256/8.

I'm committing now so that people can play with it and give any feedabck/report
any problems they find.