parse_mhs() no longer a generator function - plus other cleanups.
Because parse_mhs reads from a file, and because generators are by definition
evaluated lazily, the file must remain open at the time the generator is
evaluated.  Since this won't happen unless the caller forces the gen to
evaluate immediately [eg list( parse_mhs( .. ) )], a read on closed file
exception will be thrown - in which case, parse_mhs may as well return a
list, rather than a generator.

Other cleanups:
	* include check that fh is not a string
	* take out list() calls in nf10_sim_autosubst.py
	* remove redundant imports in mhstools.py.